Add test changes related to the last PRs to fix compilation issues

+++ src/main/java/edu/mit/dos/persistence/FileJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FileJpaRepository extends JpaRepository<File, Integer> {
 
-    File findByOid(int oid);
+    File findByOid(long oid);
 
 }
